+++ chaosazure/scripts/fill_disk.ps1
@@ -8,6 +8,6 @@ Write-Host "Filling disk with $size MB of random data for $duration seconds."
 
 $Msize = $size*1024000
 
-fsutil file createnew C:/burn $Msize
+fsutil file createnew $path $Msize
 Start-Sleep -s $duration
-rm C:/burn
+rm $path